vk_memory_manager,vk_stream_buffer: Use persistent memory maps
yuzu originally mapped and unmapped when memory had to be used from the
host. This was originally used to reduce overhead on debugging tools
(renderdoc and Nsight), but at the same time it broke compatibility with
Nsight because this is not handled properly.

After measuring recently, there isn't much of a difference in capture
and replay performance on renderdoc by avoiding persistent memory maps.

This commit replaces previous memory maps with a persistent memory map,
but the API is kept intact to be able to rollback in case it's needed.